Update datapack biome grid column assignment.
Grid column assignment for datapack biomes now use the closest vanilla bounds in Colormatic packs and the column corresponding to the raw ID in Optifine packs.

Using the closest vanilla bounds is more mod-compatible because raw IDs can shuffle, but temperature and humidity do not. Also, in general it is expected that using the bounds of a vanilla biome with similar temperature and humdity will be *close enough*.

Optifine assigns columns based on the biome's raw ID, and existing resource packs rely on this behavior, so Colormatic preserves this choice for packs using the Optifine format. Resource pack artists who wish to update their packs to the Colormatic format can replace reliance on raw IDs with explicit grid biome specifications.